Move the raw record out of each index key and into its own column.
Put that column into its own locality group, and ensure that locality group is
set when someone uses the software. Add a new method, contents(), on Sorting
that the user and fetch() can use to get said record. Update some tests to
account for the extra key per record inserted.