Add locality groups for each column provided on the SortableResult
When the user is indexing everything, we won't try to introspect the
results going in, but provide a method for the user to "optimize" the
index and add locality groups for a Set of Index's. Also, rename the
stupid member 'column' of Column to 'name' to alleviate weird looking
code.