Oof -- Finish up typed RecordValues and make the JSON parsing use them #6
Had to scrap the initial intent of using GSON's auto-magic ability to recreate
instances of a Class automatically from JSON because it was breaking the
RecordValue model I was moving towards. I could still use their JsonParser to
manually parse the JSON in the manner which I expected to receive it, which was
nicer than doing it completely by hand.